Fix integer promotion error in trig_furman. Missing begin and display commands in examples.
AVR-GCC compiler showed promotion error.
Missing begin(bitmap) in EVE_CMD_ROMFONT substitute implementation in fonts snippet.
Add a EVE_DISPLAY in the GymInterval example to enact the EVE_CMD_ROMFONT actions. (Note EVE_DISPLAY and EVE_CMD_SWAP required for locking in bitmap and font settings so they will continue to be set in subsequent display lists).

+++ examples/snippets/maths/trig_furman.h
@@ -75,8 +75,8 @@
  * Input raduis and angle in furmans
  * Returns X or Y component of vector
  */
-#define CIRC_X(radius, fur) ((radius) * sin_furman(fur) / 0x8000)
-#define CIRC_Y(radius, fur) ((radius) * cos_furman(fur) / 0x8000)
+#define CIRC_X(radius, fur) (((radius) * (uint32_t)sin_furman(fur)) / 0x8000)
+#define CIRC_Y(radius, fur) (((radius) * (uint32_t)cos_furman(fur)) / 0x8000)
 
 /* Calculate sine of furman angle
  * Input angle in furman
